+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * Add option `lib` to JuliaCall. Setting this will skip the discovery subprocess.
 * Add support for using a system image in `juliacall` that has `PythonCall` baked in.
+* Add method `pynext(x, d)` to return a default value `d` if there are no more elements.
 * Bug fixes.
 
 ## 0.9.34 (2026-05-18)